Move node syncing into an own file
